ISPN-8788 Move SingleKeyNonTxInvocationContext
Also [Non]TransactionalInvocationContextFactory
