+++ core/src/main/java/org/infinispan/factories/EmptyConstructorNamedCacheFactory.java
@@ -16,8 +16,8 @@ import org.infinispan.container.offheap.OffHeapEntryFactoryImpl;
 import org.infinispan.container.offheap.OffHeapMemoryAllocator;
 import org.infinispan.container.offheap.UnpooledOffHeapMemoryAllocator;
 import org.infinispan.context.InvocationContextFactory;
-import org.infinispan.context.NonTransactionalInvocationContextFactory;
-import org.infinispan.context.TransactionalInvocationContextFactory;
+import org.infinispan.context.impl.NonTransactionalInvocationContextFactory;
+import org.infinispan.context.impl.TransactionalInvocationContextFactory;
 import org.infinispan.distribution.L1Manager;
 import org.infinispan.distribution.RemoteValueRetrievedListener;
 import org.infinispan.distribution.TriangleOrderManager;
