+++ core/src/main/java/org/infinispan/interceptors/impl/EntryWrappingInterceptor.java
@@ -53,8 +53,8 @@ import org.infinispan.container.versioning.EntryVersion;
 import org.infinispan.container.versioning.VersionGenerator;
 import org.infinispan.context.Flag;
 import org.infinispan.context.InvocationContext;
-import org.infinispan.context.SingleKeyNonTxInvocationContext;
 import org.infinispan.context.impl.FlagBitSets;
+import org.infinispan.context.impl.SingleKeyNonTxInvocationContext;
 import org.infinispan.context.impl.TxInvocationContext;
 import org.infinispan.distribution.DistributionManager;
 import org.infinispan.distribution.ch.KeyPartitioner;
